JavaBackedType cache.
Now, to create a JavaBackedType, the method `JavaBackedType.of` must
be used which will memoize if necessary the result.

The method `JavaBackedType.of` returns a BuiltInType.UNKNOWN if it is
not possible to represent Class<> as a JavaBackedType.

The criteria if a Class<> can be represented as a JavaBackedType is 
currently implemented in `JavaBackedType.createIfAnnotated`

Added test to demonstrate currently not necessary to express properties
which are standard JavaBean compliant.
Only non-standard JavaBean compliant properties are currently needed
to be captured by JavaBackedType.